requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
appdirs==1.4.4
astroid==3.3.11
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
Bootstrap-Flask==2.5.0
bpytop==1.0.68
bs4==0.0.2
cattrs==25.1.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorgram.py==1.2.0
contourpy==1.3.2
cssselect==1.3.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.0
fake-useragent==2.2.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.4
frozenlist==1.7.0
Glances==4.3.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lsprotocol==2023.0.1
lxml==6.0.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
multidict==6.6.4
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
parse==1.20.2
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==5.9.8
pure_eval==0.2.3
pycparser==2.22
pyee==11.1.1
pygls==1.3.1
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.7
pyparsing==3.2.3
pyperclip==1.9.0
pyppeteer==2.0.0
pyquery==2.0.1
pyright==1.1.403
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
qrcode==8.2
redis==6.4.0
requests==2.32.4
requests-html==0.10.0
ruff==0.12.1
ruff-lsp==0.0.62
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spotipy==2.25.1
SQLAlchemy==2.0.44
stack-data==0.6.3
tabulate==0.9.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
ttkbootstrap==1.14.1
twilio==9.7.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
w3lib==2.3.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
windows-curses==2.4.1
wonderwords==2.2.0
wsproto==1.2.0
WTForms==3.2.1
yarl==1.20.1
zipp==3.23.0
